test(ontology): add boundary enforcement tests for upsampling
Added `test_upsampling_confidence_bounds` and `test_empty_justification_rejection`
to mathematically prove that the validators fail closed when under attack,
rejecting overconfident or evidence-free abductive leaps in the Decompression Engine.

Fixed inline import linting artifacts in `tests/contracts/test_ontology_validators.py`
and `test_ontology_payload_bounds.py` to ensure compliance with strict static analysis
in the CI/CD pipeline.
